feat: Created a simple webapp for demos / testing
## Summary

Added a simple 3 pane app with source code in the upper left, a visual
builder on the right, and a live updating result in the bottom left. I'm
using Ace Editor (with a react wrapper) for the editor panes and
`xterm.js` for the terminal emulation. A template can be produced
from the output by copying it to the clipboard (see button on UI) and
pasting it into a file of your choice (this part is definitely a WIP).

### Development

Currently can be run and developed with `yarn bundle:dev` while prod
builds can be run with `yarn bundle`. Output is put into the public dir
and care should be taken to run the commands from the root directory
of the project since there's a `rm -rf public/**` call in there and I'm
sure how it would handle relative pathing from anywhere else.